Add OpenGL rendering

+++ core/src/main/java/iskake/jgbe/core/gb/GameBoyDisplayable.java
@@ -2,6 +2,6 @@ package iskake.jgbe.core.gb;
 
 public interface GameBoyDisplayable {
     byte[] getFrame();
-    int[] getFrameMapped();
+    byte[] getFrameMapped();
     boolean canGetFrame();
 }
